+++ R/ananke-package.R
@@ -1,10 +1,16 @@
 #' @details
 #'  \tabular{ll}{
-#'   **Package:** \tab ananke \cr
-#'   **Type:** \tab Package \cr
-#'   **Version:** \tab 0.1.0 \cr
-#'   **License:** \tab GPL-3 \cr
+#'   **Version** \tab 0.2.0 \cr
+#'   **License** \tab GPL-3 \cr
+#'   **CRAN DOI** \tab \doi{10.32614/CRAN.package.ananke} \cr
+#'   **Zenodo DOI** \tab \doi{10.5281/zenodo.13236285} \cr
 #'  }
+#'
+#'  Archéosciences Bordeaux (UMR 6034)\cr
+#'  Maison de l'Archéologie\cr
+#'  Université Bordeaux Montaigne\cr
+#'  F-33607 Pessac cedex\cr
+#'  France
 #'
 #' @section Package options:
 #'  \pkg{ananke} uses the following [options()] to configure behavior:
@@ -19,30 +25,13 @@
 #'  * `ananke.verbose`: a [`logical`] scalar. Should \R report extra information
 #'    on progress? Defaults to [interactive()].
 #'
-#' @author
-#'  **Full list of authors and contributors** (alphabetic order):
-#'
-#'  \tabular{ll}{
-#'   Nicolas Frerebeau \tab *Université Bordeaux Montaigne, France* \cr
-#'  }
-#'
-#'  **Package maintainer**
-#'
-#'  Nicolas Frerebeau\cr
-#'  \email{nicolas.frerebeau@@u-bordeaux-montaigne.fr}
-#'
-#'  Archéosciences Bordeaux (UMR 6034)\cr
-#'  Maison de l'Archéologie\cr
-#'  Université Bordeaux Montaigne\cr
-#'  F-33607 Pessac cedex\cr
-#'  France
 #' @name ananke-package
 #' @aliases ananke
 #' @docType package
 #' @keywords internal
 "_PACKAGE"
 
-#' @import arkhe
 #' @import aion
-#' @importFrom methods as new setGeneric setMethod setValidity .valueClassTest
+#' @import arkhe
+#' @import methods
 NULL