+++ inst/examples/ex-f14c.R
@@ -1,5 +1,5 @@
 ## Asymmetric 14C errors (van der Plicht and Hogg 2006)
-F14C_to_BP14C(0.0052, 0.0006, asym = TRUE)
+f14c_c14(0.0052, 0.0006, asym = TRUE)
 
 ## Symmetric 14C errors (Bronk Ramsey 2008)
-F14C_to_BP14C(0.0052, 0.0006, asym = FALSE)
+f14c_c14(0.0052, 0.0006, asym = FALSE)